Update main.nf
there is an issue with a previous (--0) version of the container. This is not supported by the newest docker version anymore. We are getting: [DEPRECATION NOTICE] Docker Image Format v1 and Docker Image manifest version 2, schema 1 support is disabled by default and will be removed in an upcoming release. So the pipeline returns an error and terminates.